vendor_management_system/requirements.txt
Django==4.2.4
Django==5.0.4
djangorestframework==3.14.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
rest_framework_simplejwt==0.0.2
